Added generic implementation of BaseModelService, because all base model services need Delete And Renew functionality. Now all base model service interfaces implement IBaseModelService, and then concrete implementations extend BaseModelService. BaseModelService is abstract since, it shouldn't be created independently. Also moved specifications from Application to Domain layer, because specifications are domain logic, since they define how we can query domain entities. Moreover, Added request and response dtos for Books, but they need to be modified such that they dont use domain entities.

+++ src/Domain/Entities/Book.cs
@@ -7,7 +7,7 @@ public class Book : BaseModel
     public required string Title { get; set; }
     public string? Summary { get; set; }
     public DateTime? OriginalReleaseDate { get; set; }
-    public int? MinimumAge { get; set; }
+    public byte? MinimumAge { get; set; }
     public required Language OriginalLanguage { get; set; }
     public ICollection<Genre> Genres { get; } = [];
     public ICollection<Author> Authors { get; } = [];
